+++ apps/ssr-dev/CHANGELOG.md
@@ -1,5 +1,12 @@
 # ssr
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @dlightjs/types@0.10.4
+
 ## 0.0.1
 
 ### Patch Changes