+++ packages/core/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dlightjs/types
 
+## 0.10.4
+
+### Patch Changes
+
+- feat: default T of SubTyped
+
 ## 0.10.3
 
 ### Patch Changes