+++ packages/tools/transpiler-standalone/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @dlightjs/transpiler-standalone
 
+## 0.10.3
+
+### Patch Changes
+
+- Updated dependencies
+  - babel-preset-dlight@0.10.3
+
 ## 0.10.2
 
 ### Patch Changes